requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Automat==0.8.0
blinker==1.4
Brotli==1.0.9
cchardet==2.1.7
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
ciso8601==2.2.0
colorama==0.4.3
colorlog==6.6.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
ed25519==1.5
entrypoints==0.3
frozenlist==1.3.0
hikari==2.0.0.dev105
hikari-lightbulb==2.1.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==5.2.0
netifaces==0.10.4
numpy==1.22.1
oauthlib==3.1.0
pexpect==4.6.0
protobuf==3.6.1
pure25519==0.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycares==4.1.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-dotenv==0.19.2
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
urllib3==1.25.8
uvloop==0.16.0
wadllib==1.3.3
yarl==1.7.2
zipp==1.0.0
